requirement.txt
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
execnet==2.1.1
fake-headers==1.0.2
Faker==25.3.0
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
idna==3.7
iniconfig==2.0.0
lxml==5.2.2
markdown-it-py==3.0.0
mdurl==0.1.2
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
Pygments==2.18.0
PySocks==1.7.1
pyTelegramBotAPI==4.18.1
pytest==8.2.2
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
vk-api==11.9.9
webdriver-manager==4.0.1
webencodings==0.5.1
wsproto==1.2.0
